backend/requirements.txt
fastapi[standard]==0.115.12
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.0
aiosqlite==0.21.0
pytest==8.3.5
google-api-python-client==2.169.0
google-auth-httplib2==0.2.0 
google-auth-oauthlib==1.2.2
faker==37.1.0
alembic==1.15.2 
fastapi-sso==0.18.0 
ruff==0.11.10